Move choosing of outbox folders to downstream provider.
